Moves sqlalchemy 1.x functionality into a "contrib".
