+++ .pre-commit-config.yaml
@@ -109,7 +109,7 @@ repos:
             pyyaml,
             redis,
             rich,
-            sqlalchemy<2.0.0,
+            sqlalchemy>2.0,
             starlette,
             structlog,
             tortoise_orm,
@@ -152,8 +152,7 @@ repos:
             python-jose,
             pyyaml,
             rich,
-            sqlalchemy<2.0.0,
-            sqlalchemy2-stubs,
+            sqlalchemy>2.0,
             starlette,
             structlog,
             tortoise-orm,
@@ -199,8 +198,7 @@ repos:
             python-jose,
             pyyaml,
             rich,
-            sqlalchemy<2.0.0,
-            sqlalchemy2-stubs,
+            sqlalchemy>2.0,
             starlette,
             structlog,
             tortoise-orm,
