+++ docs/examples/data_transfer_objects/dto_auto_conversion.py
@@ -4,9 +4,9 @@ from sqlalchemy import Column, Float, Integer, String
 from sqlalchemy.orm import Mapped, declarative_base
 
 from starlite import Starlite, get
+from starlite.contrib.sqlalchemy_1.plugin import SQLAlchemyPlugin
 from starlite.dto import DTOFactory
 from starlite.exceptions import HTTPException
-from starlite.plugins.sql_alchemy import SQLAlchemyPlugin
 from starlite.status_codes import HTTP_404_NOT_FOUND
 
 sqlalchemy_plugin = SQLAlchemyPlugin()
