+++ docs/examples/data_transfer_objects/dto_basic.py
@@ -1,8 +1,8 @@
 from sqlalchemy import Column, Float, Integer, String
 from sqlalchemy.orm import Mapped, declarative_base
 
+from starlite.contrib.sqlalchemy_1.plugin import SQLAlchemyPlugin
 from starlite.dto import DTOFactory
-from starlite.plugins.sql_alchemy import SQLAlchemyPlugin
 
 dto_factory = DTOFactory(plugins=[SQLAlchemyPlugin()])
