+++ docs/examples/plugins/sqlalchemy_plugin/sqlalchemy_relationships.py
@@ -5,7 +5,8 @@ from sqlalchemy.orm import Mapped, Session, declarative_base, relationship
 
 from starlite import Starlite, get
 from starlite.exceptions import HTTPException
-from starlite.plugins.sql_alchemy import SQLAlchemyConfig, SQLAlchemyPlugin
+from starlite.contrib.sqlalchemy_1.config import SQLAlchemyConfig
+from starlite.contrib.sqlalchemy_1.plugin import SQLAlchemyPlugin
 from starlite.status_codes import HTTP_404_NOT_FOUND
 
 engine = create_engine("sqlite+pysqlite://")
