+++ docs/examples/tests/application_hooks/test_application_after_exception_hook.py
@@ -2,7 +2,6 @@ import logging
 from typing import TYPE_CHECKING
 
 from examples.application_hooks import after_exception_hook
-
 from starlite.testing import TestClient
 
 if TYPE_CHECKING:
