+++ docs/examples/tests/application_hooks/test_application_before_send.py
@@ -1,5 +1,4 @@
 from examples.application_hooks import before_send_hook
-
 from starlite.status_codes import HTTP_200_OK
 from starlite.testing import TestClient
