+++ docs/examples/tests/application_hooks/test_application_shutdown_hooks.py
@@ -2,7 +2,6 @@ import logging
 from typing import TYPE_CHECKING
 
 from examples.application_hooks import shutdown_hooks
-
 from starlite.testing import TestClient
 
 if TYPE_CHECKING:
