+++ docs/examples/tests/dependency_injection/test_dependency_default_value_no_dependency_fn.py
@@ -1,5 +1,4 @@
 from examples.dependency_injection import dependency_default_value_no_dependency_fn
-
 from starlite.status_codes import HTTP_200_OK
 from starlite.testing import TestClient
