+++ docs/examples/tests/dependency_injection/test_dependency_skip_validation.py
@@ -1,5 +1,4 @@
 from examples.dependency_injection import dependency_skip_validation
-
 from starlite.status_codes import HTTP_200_OK
 from starlite.testing import TestClient
