+++ docs/examples/tests/openapi/test_customize_pydantic_model_name.py
@@ -1,5 +1,4 @@
 from examples.openapi import customize_pydantic_model_name
-
 from starlite.testing import TestClient
