+++ docs/examples/tests/test_exceptions.py
@@ -3,7 +3,6 @@ from examples.exceptions import (
     override_default_handler,
     per_exception_handlers,
 )
-
 from starlite.testing import TestClient
