+++ docs/examples/tests/test_hello_world.py
@@ -1,5 +1,4 @@
 from examples import hello_world
-
 from starlite.status_codes import HTTP_200_OK
 from starlite.testing import TestClient
