+++ docs/examples/tests/test_startup_and_shutdown.py
@@ -2,7 +2,6 @@ from typing import TYPE_CHECKING
 from unittest.mock import AsyncMock, MagicMock
 
 from examples import startup_and_shutdown
-
 from starlite import get
 from starlite.datastructures import State
 from starlite.testing import TestClient
