+++ mypy.ini
@@ -12,6 +12,12 @@ disallow_any_generics = False
 implicit_reexport = False
 show_error_codes = True
 
+[mypy-starlite.contrib.sqlalchemy_1.*]
+ignore_errors = True
+
+[mypy-tests.contrib.sqlalchemy_1.*]
+ignore_errors = True
+
 [mypy-starlite.plugins.sql_alchemy.plugin]
 disable_error_code = dict-item
