--- starlite/plugins/sql_alchemy/__init__.py
@@ -1,4 +1,0 @@
-from .config import SQLAlchemyConfig, SQLAlchemyEngineConfig, SQLAlchemySessionConfig
-from .plugin import SQLAlchemyPlugin
-
-__all__ = ("SQLAlchemyPlugin", "SQLAlchemyConfig", "SQLAlchemyEngineConfig", "SQLAlchemySessionConfig")
