+++ tests/contrib/sqlalchemy_1/conftest.py
@@ -1,0 +1,9 @@
+from __future__ import annotations
+
+try:
+    import sqlalchemy
+except ImportError:
+    collect_ignore_glob = ["*"]
+else:
+    if sqlalchemy.__version__.startswith("1."):
+        collect_ignore_glob = ["*"]
