+++ tests/contrib/sqlalchemy_1/sql_alchemy_plugin/test_sql_alchemy_config.py
@@ -7,15 +7,15 @@ from sqlalchemy.orm import Session, sessionmaker
 
 from starlite import Starlite, get
 from starlite.config.logging import LoggingConfig
-from starlite.exceptions import ImproperlyConfiguredException
-from starlite.plugins.sql_alchemy import SQLAlchemyPlugin
-from starlite.plugins.sql_alchemy.config import (
+from starlite.contrib.sqlalchemy_1.config import (
     SESSION_SCOPE_KEY,
     SESSION_TERMINUS_ASGI_EVENTS,
     SQLAlchemyConfig,
     SQLAlchemyEngineConfig,
     serializer,
 )
+from starlite.contrib.sqlalchemy_1.plugin import SQLAlchemyPlugin
+from starlite.exceptions import ImproperlyConfiguredException
 from starlite.status_codes import HTTP_200_OK
 from starlite.testing import RequestFactory, create_test_client
 from starlite.types import Scope
