+++ tests/contrib/sqlalchemy_1/sql_alchemy_plugin/test_sql_alchemy_models_openapi_generation.py
@@ -1,6 +1,6 @@
 from starlite import Starlite, get, post
-from starlite.plugins.sql_alchemy import SQLAlchemyPlugin
-from tests.plugins.sql_alchemy_plugin import User
+from starlite.contrib.sqlalchemy_1.plugin import SQLAlchemyPlugin
+from tests.contrib.sqlalchemy_1.sql_alchemy_plugin import User
 
 
 @get(path="/user")
