+++ tests/contrib/sqlalchemy_1/sql_alchemy_plugin/test_sql_alchemy_table_types.py
@@ -54,10 +54,10 @@ from sqlalchemy.sql.functions import now
 # TupleType not a sqlalchemy2-stubs top-level import
 from sqlalchemy.sql.sqltypes import TupleType
 
+from starlite.contrib.sqlalchemy_1.plugin import SQLAlchemyPlugin
 from starlite.exceptions import ImproperlyConfiguredException
-from starlite.plugins.sql_alchemy import SQLAlchemyPlugin
 from tests import Species
-from tests.plugins.sql_alchemy_plugin import SQLAlchemyBase
+from tests.contrib.sqlalchemy_1.sql_alchemy_plugin import SQLAlchemyBase
 
 plugin = SQLAlchemyPlugin()
