+++ tests/dto_factory/test_dto_factory_integration.py
@@ -6,12 +6,12 @@ from pydantic import BaseModel
 from pydantic_factories import ModelFactory
 
 from starlite import post
+from starlite.contrib.sqlalchemy_1.plugin import SQLAlchemyPlugin
 from starlite.dto import DTOFactory
-from starlite.plugins.sql_alchemy import SQLAlchemyPlugin
 from starlite.status_codes import HTTP_201_CREATED
 from starlite.testing import create_test_client
 from tests import Person, TypedDictPerson, VanillaDataClassPerson
-from tests.plugins.sql_alchemy_plugin import Pet, WildAnimal
+from tests.contrib.sqlalchemy_1.sql_alchemy_plugin import Pet, WildAnimal
 
 
 @pytest.mark.parametrize(
